minor bugs
