+++ graphlearn/utils/draw.py
@@ -4,11 +4,33 @@ import networkx as nx
 import numpy as np
 from scipy.optimize import curve_fit
 from collections import defaultdict
-#from graphlearn.utils import calc_stats_from_grammar
 import logging
 import copy
 
+
 logger = logging.getLogger(__name__)
+
+
+
+
+
+
+
+
+def calc_stats_from_grammar(grammar):
+    count_corehashes = defaultdict(int)
+    count_interfacehashes = defaultdict(int)
+    corecounter = defaultdict(int)
+    intercounter = defaultdict(int)
+    for ih in grammar.keys():
+        for ch in grammar[ih].keys():
+            # go over all the combos
+            count_corehashes[ch] += 1
+            count_interfacehashes[ih] += 1
+            count = grammar[ih][ch].count
+            corecounter[ch] += count
+            intercounter[ih] += count
+    return count_corehashes, count_interfacehashes, corecounter, intercounter
 
 '''
     functions to draw graphs:
